Cycle 156b: TreeGrowthConfig — ConsciousnessTree hardcoded thresholds centralized (D35)
Created TreeGrowthConfig struct with 15 centralized threshold fields:
fruit_growth_health, max_growth_violations, fruit_quality_threshold,
action_severity_threshold, neutral_health, min_branch_health,
soil_health_critical, root_health_high, branch_maturity_low,
phi_minimum, phi_warmup_cycles, constitution_rules/experiences/floor,
constitution_tree_growth_rules, constitution_absorption_rules.

Added config field to ConsciousnessTree, kept backward-compatible new()
signature, wired TreeGrowthConfig::default() in constructor.
Replaced all 15 hardcoded thresholds in run_growth_cycle() and
scan_vulnerabilities() with self.config.* references.

CVSS tiers, evidence credibility tiers, evidence type tiers are domain
classification boundaries — kept inline as appropriate.

SelfTest: 5/5 pass (6682 total, 0 failed). Build: 0 errors (71 warnings).

+++ neotrix-core/src/core/nt_core_gwt/consensus_resonance.rs
@@ -47,10 +47,10 @@ impl ConsensusResonanceBridge {
                 } else {
                     0
                 };
-                let modulated = (base as u32).saturating_add(modulation).min(6);
+                let modulated = base.saturating_add(modulation).min(6);
                 strengths[i][j] = modulated;
                 total_strength += modulated as u64;
-                if modulated >= RESONANCE_THRESHOLD as u32 && i != j {
+                if modulated >= RESONANCE_THRESHOLD && i != j {
                     in_resonance += 1;
                 }
             }
